+++ rtos/net/net_tcp.c
@@ -583,7 +583,7 @@ static void tcp_rtx_start(TCP_PORT *port)
 static void tcp_rtx_start_timeout(TCP_PORT *port, uint32_t timeout)
 {
     /* Set required next timeout. */
-    port->rtx_data.suspend.timeout = current_system_tick() + timeout;
+    port->rtx_data.suspend.timeout = (uint32_t)(current_system_tick() + timeout);
 
     /* Networking condition data has been updated. */
     net_condition_updated();
@@ -632,7 +632,7 @@ static void tcp_rtx_stop(TCP_PORT *port)
 static void tcp_rtx_callback(void *data)
 {
     TCP_PORT *port = (TCP_PORT *)data;
-    uint64_t next_timeout;
+    uint32_t next_timeout;
 
     /* Get lock for this port. */
     if (fd_get_lock((FD)port) == SUCCESS)
@@ -691,7 +691,7 @@ static void tcp_rtx_callback(void *data)
                 }
 
                 /* SET (ExpBoff × RTO, REXMT) */
-                next_timeout = port->expboff * TCP_RTO;
+                next_timeout = (uint32_t)(port->expboff * TCP_RTO);
 
                 /* SSthresh := max (SND.WND/2, 2) */
                 port->ssthresh = (uint16_t)MAX((port->snd_wnd/2), 2);
@@ -718,7 +718,7 @@ static void tcp_rtx_callback(void *data)
         if (next_timeout != MAX_WAIT)
         {
             /* Schedule next time out. */
-            port->rtx_data.suspend.timeout = current_system_tick() + next_timeout;
+            port->rtx_data.suspend.timeout = (uint32_t)(current_system_tick() + next_timeout);
         }
         else
         {
